Updated quintile function to handle low N
There was a bug where if the vector of values didn't have enough variability to create 6 unique breakpoints, the cut() function would throw an error. We are now passing the unique cutoff points to the cut() function, and replacing whatever the highest quintile ends up being with 5. So if there are only 4 unique cutoff points, resulting in 3 groups, the labels will be 1, 2, 5.